LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-multicol-colspan-001-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-offset-x="10" data-offset-y="0" data-expected-width="40" data-expected-height="20"></div>
width expected 40 but got 0
FAIL .target 2 assert_equals:
<div class="target" data-offset-x="10" data-offset-y="20" data-expected-width="40" data-expected-height="20"></div>
width expected 40 but got 0
PASS .target 3
FAIL .target 4 assert_equals:
<div class="target" data-offset-x="10" data-offset-y="20" data-expected-width="40" data-expected-height="20"></div>
width expected 40 but got 0
FAIL .target 5 assert_equals:
<div class="target" data-offset-x="10" data-offset-y="30" data-expected-width="40" data-expected-height="20"></div>
width expected 40 but got 0
